refactor(dash/types): rename package dashtypes to types
