+++ dash/types/nodeid_resolver.go
@@ -1,4 +1,4 @@
-package dashtypes
+package types
 
 import (
 	"fmt"
