+++ dash/types/validator_address.go
@@ -1,4 +1,4 @@
-package dashtypes
+package types
 
 import (
 	"errors"
