+++ dash/types/validator_address_test.go
@@ -1,4 +1,4 @@
-package dashtypes
+package types
 
 import (
 	"testing"
